refactor: comment 파일 분리
